+++ src/settings.jl
@@ -125,6 +125,18 @@ $(TYPEDFIELDS)
     rel_nose_mass         = 0
     "mass of the top particle relative to the sum of top and side particles"
     rel_top_mass          = 0
+    
+    # model KPS4_3L
+    "the radius of the circle shape on which the kite lies, viewed from the front [m]"
+    radius::Float64 = 10.0
+    "the distance from point the center bridle connection point of the middle line to the kite [m]"
+    bridle_center_distance::Float64 = 2.0
+    "the cord length of the kite in the middle [m]"
+    middle_length::Float64 = 2.0
+    "the cord length of the kite at the tips [m]"
+    tip_length::Float64 = 1.0
+    "the distance between the left and right steering bridle line connections on the kite that are closest to eachother [m]"
+    min_steering_line_distance::Float64 = 4.0 
 
     "bridle line diameter                  [mm]"
     d_line                = 0